assessments/requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20

# Document Processing
beautifulsoup4==4.12.3
lxml==5.1.0
pypdf==4.0.1
python-docx==1.1.0

# Embeddings & Vector Search
openai==1.12.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
chromadb==0.4.22

# Sparse Retrieval
rank-bm25==0.2.2

# Reranking
transformers==4.37.2
torch==2.2.0

# Utilities
tiktoken==0.5.2
numpy==1.26.3
pandas==2.2.0
tqdm==4.66.1

# Evaluation
ragas==0.1.4
scikit-learn==1.4.0

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Development
jupyter==1.0.0
ipykernel==6.29.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Monitoring
python-dotenv==1.0.0
loguru==0.7.2

# Additional utilities
requests==2.31.0
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.1